requirements.txt
hydra-core>=1.3.2
matplotlib
numpy>=1.23.4
omegaconf
opencv_python
pandas
PyYAML
scipy>=1.10.0
tqdm
ultralytics
torch
motrack-motion>=0.2.4
pydantic>=2.12.5
